area-66-implement-user-registration-and-login :sparkles: Add login and register for users
