--- server/src/typeorm/entities/UserEntity.ts
@@ -1,19 +1,0 @@
-import { Column, Entity, PrimaryGeneratedColumn } from 'typeorm';
-
-@Entity({ name: 'users' })
-export class UserEntity {
-  @PrimaryGeneratedColumn()
-  id: number;
-
-  @Column()
-  email: string;
-
-  @Column()
-  username: string;
-
-  @Column()
-  password: string;
-
-  // @Column()
-  // accessToken: string;
-}
