--- server/src/users/dto/login-user.dto.ts
@@ -1,4 +1,0 @@
-export class LoginUserDto {
-    email: string;
-    password: string;
-}
