+++ server/src/users/dto/user-login.dto.ts
@@ -1,0 +1,4 @@
+export class LoginUserDto {
+    readonly username: string;
+    readonly password: string;
+}
