+++ server/src/users/dto/user.create.dto.ts
@@ -1,5 +1,5 @@
 export class CreateUserDto {
-    email: string;
     username: string;
     password: string;
+    email: string;
 }
